+++ frontend/package.json
@@ -29,7 +29,6 @@
     "react-hot-toast": "^2.4.1",
     "react-hotkeys-hook": "^4.4.1",
     "react-markdown": "^8.0.7",
-    "react-resizable": "^3.0.5",
     "react-router-dom": "^6.15.0",
     "react-syntax-highlighter": "^15.5.0",
     "react-virtualized-auto-sizer": "^1.0.20",
@@ -53,7 +52,6 @@
     "@types/draft-js": "^0.11.10",
     "@types/lodash": "^4.14.197",
     "@types/node": "^20.5.7",
-    "@types/react-resizable": "^3.0.4",
     "@types/react-syntax-highlighter": "^15.5.7",
     "@types/react-window-infinite-loader": "^1.0.6",
     "@types/uuid": "^9.0.3",
