+++ libs/components/package.json
@@ -35,6 +35,7 @@
     "formik": "^2.4.3",
     "mui-chips-input": "^2.1.3",
     "react-markdown": "^8.0.7",
+    "react-resizable": "^3.0.5",
     "react-syntax-highlighter": "^15.5.0",
     "remark-gfm": "^3.0.1",
     "swr": "^2.2.2",
@@ -42,6 +43,7 @@
     "yup": "^1.2.0"
   },
   "devDependencies": {
+    "@types/react-resizable": "^3.0.4",
     "@types/react-syntax-highlighter": "^15.5.7",
     "@vitejs/plugin-react-swc": "^3.3.2",
     "tsup": "^7.2.0",
