+++ libs/components/src/elements/index.ts
@@ -1,5 +1,8 @@
 export { AudioElement } from './Audio';
 export { AvatarElement } from './Avatar';
+export { Element } from './Element';
+export { ElementSideView } from './ElementSideView';
+export { ElementView } from './ElementView';
 export { FileElement } from './File';
 export { FrameElement } from './Frame';
 export { ImageElement } from './Image';
